+++ apps/test-app/src/app/app.config.ts
@@ -1,9 +1,10 @@
-import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
-import { provideRouter } from '@angular/router';
-import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
+import { ApplicationConfig, provideZoneChangeDetection } from "@angular/core";
+import { provideRouter } from "@angular/router";
+import { provideAnimationsAsync } from "@angular/platform-browser/animations/async";
 import { provideSsvCommandOptions } from "@ssv/ngx.command";
+import { provideSsvUxViewportOptions } from "@ssv/ngx.ux";
 
-import { appRoutes } from './app.routes';
+import { appRoutes } from "./app.routes";
 
 export const appConfig: ApplicationConfig = {
 	providers: [
@@ -15,5 +16,21 @@ export const appConfig: ApplicationConfig = {
 			executingCssClass: "is-busy",
 			hasDisabledDelay: false
 		}),
+
+		provideSsvUxViewportOptions({
+			// breakpoints: {
+			// 	small: 1000,
+			// }
+		},
+		//  withViewportSsrDevice(() => DeviceType.mobile)
+		),
+		// provideSsvUxViewportOptions(defaults => {
+		// 	return {
+		// 		breakpoints: {
+		// 			...defaults.breakpoints,
+		// 			small: 1000,
+		// 		}
+		// 	};
+		// }),
 	],
 };