+++ libs/ngx.ux/src/index.ts
@@ -1,0 +1,4 @@
+export { SsvUxModule } from "./ux.module";
+export * from "./viewport/index";
+
+export { VERSION } from "./version";